requirements.txt
asgiref==3.2.5
Django==3.0.4
django-rest-registration==0.5.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
typing==3.7.4.1
